Stop stored hot water crash + device naming fix
A quick and dirty fix to stop the service crashing if you have fixed hot water. It doesn't actually give you the ability to control the hot water, it just ignores it completely.
Also, each thermostat was showing up with the same device name, so I'd no idea which room each stat was in. This has been fixed too.